production/packages/utils/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = "Utils for the Discord-RAG project"
authors = ["antoinelrnld <38052726+antoinelrnld@users.noreply.github.com>"]
license = "MIT"
packages = [{ include = "utils", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
langchain-experimental = "0.3.4"
langchain-community = "0.3.13"
langchain-core = "0.3.28"
langchain-redis = "0.1.1"
google-generativeai = ">=0.8.0"
motor = "3.6.0"
tqdm = "^4.66.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
